chore: clean up console logs and keep only database in docker-compose
