+++ src/main.ts
@@ -11,7 +11,6 @@ async function bootstrap() {
     transform: false, //transforma os tipos primitivos (string, number, boolean) conforme definido no DTO
   }
 ))
-console.log('porta', process.env.PORT);
   await app.listen(process.env.PORT ?? 3000);
   
 }
