+++ src/user/dto/create-user.dto.ts
@@ -1,0 +1,16 @@
+import { IsEmail, IsNotEmpty, IsString, MinLength } from "class-validator";
+
+export class CreateUserDto {
+  @IsString()
+  @IsNotEmpty()
+  name: string;
+
+  @IsEmail()
+  @IsNotEmpty()
+  email: string;
+
+  @IsString()
+  @MinLength(6)
+  @IsNotEmpty()
+  password: string;
+}
